Move SimulateV1 rpc function to api.go
